`eosio_types.mjs` now exported from root directory.
